# perf: use Sets instead of RegExes in normalizeLegalName