--- a/src/lib/geoip/legal-name-normalization.ts
+++ b/src/lib/geoip/legal-name-normalization.ts
@@ -18,6 +18,7 @@ const ADDITIONAL_LEGAL_FORMS = [
 	{ name: 'Joint Stock Company', abbr: [ 'JSC' ] }, // Russia
 	{ name: 'Public Company Limited', abbr: [ 'PCL' ] }, // Thailand
 	{ name: 'Liability Company', abbr: [ 'LC' ] }, // Vietnam
+	{ name: 'spolecnost s Rucenim Omezenym', abbr: [ 'SRO' ] }, // Slovakia
 ];
 
 // Some languages have custom transliteration rules.
@@ -30,6 +31,11 @@ const PREFIX_USING_COUNTRIES = [ 'RO', 'LV', 'LT', 'EE', 'RU', 'UA', 'BY', 'MD',
 const normalizedCache = new LRUCache<string, string>({
 	max: 10000,
 });
+
+const allNamesSet = new Set<string>();
+const allAbbrsSet = new Set<string>();
+const prefixNamesSet: Set<string> = new Set();
+const prefixAbbrsSet: Set<string> = new Set();
 
 let legalSuffixNamesPattern: RegExp;
 let legalSuffixAbbrsPattern: RegExp;
@@ -56,6 +62,72 @@ type CsvLegalFormRow = {
 };
 
 export const normalizeLegalName = (name: string) => {
+	if (!legalSuffixNamesPattern || !legalSuffixAbbrsPattern || !legalPrefixNamesPattern || !legalPrefixAbbrsPattern) {
+		throw new Error('Legal name normalization is not initialized.');
+	}
+
+	const normalized = name.trim()
+		// Normalize "trading as" names, e.g., "Matteo Martelloni trading as DELUXHOST" => "DELUXHOST"
+		.split(/\s+trading as\s+/i).at(-1)!
+		// Clean up any double spaces.
+		.replace(/\s+/g, ' ')
+		// Add missing space after commas, e.g., "Hangzhou Alibaba Advertising Co.,Ltd." => "Hangzhou Alibaba Advertising"
+		.replace(/,(?=\S)/g, ', ');
+
+	const words = normalized.split(/\s+/);
+
+	// Searching prefix.
+	const firstWord = words[0]!.toLowerCase().replace(/\./g, '');
+
+	if (prefixNamesSet.has(firstWord) || prefixAbbrsSet.has(firstWord)) {
+		words.splice(0, 1);
+	}
+
+	// Searching suffix. Starting from the 2nd word, try to remove the longest possible combination.
+	for (let i = 1; i < words.length; i++) {
+		const combination = words.slice(i).join('').toLowerCase()
+			// Suffixes Sets don't include dots, parentheses, commas and spaces, so we are removing them from the combination.
+			.replace(/[.,()]/g, '');
+
+		// Remove the last word if it's a hyphen or an ampersand and proceed with suffix search.
+		if (([ '-', '&' ].includes(words[i]!) && i === words.length - 1)) {
+			words.splice(i);
+			i = 0;
+			continue;
+		}
+
+		if (
+			allNamesSet.has(combination)
+			|| allAbbrsSet.has(combination)
+		) {
+			words.splice(i);
+
+			if (
+				// If prev word ends with one of the following, proceed with suffix search.
+				words[i - 1]!.endsWith('.')
+				|| words[i - 1]!.endsWith('.,')
+				|| words[i - 1]!.endsWith(')')
+				|| words[i - 1]!.endsWith('),')
+				// Move to the next iteration to remove hyphen or ampersand there.
+				|| [ '-', '&' ].includes(words[i - 1]!)
+			) {
+				i = 0;
+			} else {
+				break;
+			}
+		}
+	}
+
+	const result = words.join(' ')
+		// Remove trailing commas and spaces after suffix removal.
+		.replace(/\s*,\s*$/, '')
+		// Remove wrapping quotes that are often used with prefixes.
+		.replace(/^"(.*)"$/, '$1');
+
+	return result;
+};
+
+export const normalizeLegalName2 = (name: string) => {
 	if (normalizedCache.has(name)) {
 		return normalizedCache.get(name)!;
 	}
@@ -87,6 +159,10 @@ export const populateLegalNames = async () => {
 	const { allForms, prefixForms } = await readLegalFormsFile();
 	const { names: allNames, abbrs: allAbbrs } = await collectLegalForms(allForms);
 	const { names: prefixNames, abbrs: prefixAbbrs } = await collectLegalForms(prefixForms, 3);
+	allNames.forEach(name => allNamesSet.add(name));
+	allAbbrs.forEach(abbr => allAbbrsSet.add(abbr));
+	prefixNames.forEach(name => prefixNamesSet.add(name));
+	prefixAbbrs.forEach(abbr => prefixAbbrsSet.add(abbr));
 	({ namesPattern: legalSuffixNamesPattern, abbrsPattern: legalSuffixAbbrsPattern } = buildSuffixPatterns(allNames, allAbbrs));
 	({ namesPattern: legalPrefixNamesPattern, abbrsPattern: legalPrefixAbbrsPattern } = buildPrefixPatterns(prefixNames, prefixAbbrs));
 };
@@ -159,7 +235,7 @@ async function collectLegalForms (legalFormsData: CsvLegalFormRow[], minSynthesi
 
 		abbrLocal.split(';').forEach((abbr) => {
 			if (abbr.trim()) {
-				legalFormsAbbr.add(toAscii(abbr.trim()));
+				legalFormsAbbr.add(toAscii(abbr.trim()).replace(/[., ()]/g, ''));
 			}
 		});
 
@@ -167,20 +243,20 @@ async function collectLegalForms (legalFormsData: CsvLegalFormRow[], minSynthesi
 
 		abbrTransliterated.split(';').forEach((abbr) => {
 			if (abbr.trim()) {
-				legalFormsAbbr.add(toAscii(abbr.trim()));
+				legalFormsAbbr.add(toAscii(abbr.trim()).replace(/[., ()]/g, ''));
 			}
 		});
 
 		const nameLocal = row.entityLegalFormNameLocal.trim().toLowerCase();
 
 		if (nameLocal) {
-			legalFormsName.add(toAscii(nameLocal));
+			legalFormsName.add(toAscii(nameLocal).replace(/[., ()]/g, ''));
 
 			if (!abbrLocal) {
 				const abbrParts = nameLocal.split(' ').filter(is.truthy);
 
 				if (abbrParts.length >= minSynthesizedAbbreviationLength) {
-					legalFormsAbbr.add(toAscii(abbrParts.map(w => `${w[0]}.`).join('')));
+					legalFormsAbbr.add(toAscii(abbrParts.map(w => `${w[0]}`).join('')).replace(/[., ()]/g, ''));
 				}
 			}
 		}
@@ -188,13 +264,13 @@ async function collectLegalForms (legalFormsData: CsvLegalFormRow[], minSynthesi
 		const nameTransliterated = row.entityLegalFormNameTransliterated.trim().toLowerCase();
 
 		if (nameTransliterated) {
-			legalFormsName.add(toAscii(nameTransliterated.trim()));
+			legalFormsName.add(toAscii(nameTransliterated.trim()).replace(/[., ()]/g, ''));
 		}
 	});
 
 	ADDITIONAL_LEGAL_FORMS.forEach(({ name, abbr }) => {
-		legalFormsName.add(ascii(name).toLowerCase());
-		abbr.forEach(a => legalFormsAbbr.add(ascii(a).toLowerCase()));
+		legalFormsName.add(ascii(name).toLowerCase().replace(/[., ()]/g, ''));
+		abbr.forEach(a => legalFormsAbbr.add(ascii(a).toLowerCase().replace(/[., ()]/g, '')));
 	});
 
 	// Convert to array and sort by length (longest first) to avoid partial matches