--- a/src/lib/geoip/legal-name-normalization.ts
+++ b/src/lib/geoip/legal-name-normalization.ts
@@ -4,7 +4,6 @@ import csvParser from 'csv-parser';
 import { LRUCache } from 'lru-cache';
 import transliterate from '@sindresorhus/transliterate';
 import is from '@sindresorhus/is';
-import _ from 'lodash';
 
 // See https://github.com/jsdelivr/globalping/issues/383
 // The CSV file is from https://www.gleif.org/en/lei-data/code-lists/iso-20275-entity-legal-forms-code-list
@@ -18,7 +17,7 @@ const ADDITIONAL_LEGAL_FORMS = [
 	{ name: 'Joint Stock Company', abbr: [ 'JSC' ] }, // Russia
 	{ name: 'Public Company Limited', abbr: [ 'PCL' ] }, // Thailand
 	{ name: 'Liability Company', abbr: [ 'LC' ] }, // Vietnam
-	{ name: 'spolecnost s Rucenim Omezenym', abbr: [ 'SRO' ] }, // Slovakia
+	{ name: 'Spolecnost s Rucenim Omezenym', abbr: [ 'SRO' ] }, // Slovakia
 ];
 
 // Some languages have custom transliteration rules.
@@ -36,11 +35,6 @@ const allNamesSet = new Set<string>();
 const allAbbrsSet = new Set<string>();
 const prefixNamesSet: Set<string> = new Set();
 const prefixAbbrsSet: Set<string> = new Set();
-
-let legalSuffixNamesPattern: RegExp;
-let legalSuffixAbbrsPattern: RegExp;
-let legalPrefixNamesPattern: RegExp;
-let legalPrefixAbbrsPattern: RegExp;
 
 type CsvLegalFormRow = {
 	elfCode: string;
@@ -62,7 +56,11 @@ type CsvLegalFormRow = {
 };
 
 export const normalizeLegalName = (name: string) => {
-	if (!legalSuffixNamesPattern || !legalSuffixAbbrsPattern || !legalPrefixNamesPattern || !legalPrefixAbbrsPattern) {
+	if (normalizedCache.has(name)) {
+		return normalizedCache.get(name)!;
+	}
+
+	if (!allNamesSet.size || !allAbbrsSet.size || !prefixNamesSet.size || !prefixAbbrsSet.size) {
 		throw new Error('Legal name normalization is not initialized.');
 	}
 
@@ -74,10 +72,10 @@ export const normalizeLegalName = (name: string) => {
 		// Add missing space after commas, e.g., "Hangzhou Alibaba Advertising Co.,Ltd." => "Hangzhou Alibaba Advertising"
 		.replace(/,(?=\S)/g, ', ');
 
-	const words = normalized.split(/\s+/);
+	const words = normalized.split(' ');
 
 	// Searching prefix.
-	const firstWord = words[0]!.toLowerCase().replace(/\./g, '');
+	const firstWord = words[0]!.toLowerCase().replace(/[.,()]/g, '');
 
 	if (prefixNamesSet.has(firstWord) || prefixAbbrsSet.has(firstWord)) {
 		words.splice(0, 1);
@@ -124,35 +122,9 @@ export const normalizeLegalName = (name: string) => {
 		// Remove wrapping quotes that are often used with prefixes.
 		.replace(/^"(.*)"$/, '$1');
 
+	normalizedCache.set(name, result);
+
 	return result;
-};
-
-export const normalizeLegalName2 = (name: string) => {
-	if (normalizedCache.has(name)) {
-		return normalizedCache.get(name)!;
-	}
-
-	if (!legalSuffixNamesPattern || !legalSuffixAbbrsPattern || !legalPrefixNamesPattern || !legalPrefixAbbrsPattern) {
-		throw new Error('Legal name normalization is not initialized.');
-	}
-
-	const normalized = name.trim()
-		// Normalize "trading as" names, e.g., "Matteo Martelloni trading as DELUXHOST" => "DELUXHOST"
-		.split(/\s+trading as\s+/i).at(-1)!
-		// Apply the main cleanup patterns.
-		.replace(legalSuffixNamesPattern, '').trim()
-		.replace(legalSuffixAbbrsPattern, '').trim()
-		.replace(legalPrefixNamesPattern, '').trim()
-		.replace(legalPrefixAbbrsPattern, '').trim()
-		// Clean up any double spaces.
-		.replace(/\s+/g, ' ')
-		// Remove wrapping quotes that are often used with prefixes.
-		.replace(/^"(.*)"$/, '$1')
-		// Remove trailing commas and spaces after suffix removal.
-		.replace(/\s*,\s*$/, '');
-
-	normalizedCache.set(name, normalized);
-	return normalized;
 };
 
 export const populateLegalNames = async () => {
@@ -163,64 +135,7 @@ export const populateLegalNames = async () => {
 	allAbbrs.forEach(abbr => allAbbrsSet.add(abbr));
 	prefixNames.forEach(name => prefixNamesSet.add(name));
 	prefixAbbrs.forEach(abbr => prefixAbbrsSet.add(abbr));
-	({ namesPattern: legalSuffixNamesPattern, abbrsPattern: legalSuffixAbbrsPattern } = buildSuffixPatterns(allNames, allAbbrs));
-	({ namesPattern: legalPrefixNamesPattern, abbrsPattern: legalPrefixAbbrsPattern } = buildPrefixPatterns(prefixNames, prefixAbbrs));
 };
-
-function buildSuffixPatterns (names: string[], abbrs: string[]) {
-	return buildPatterns(names, abbrs, (patterns: string[]) => {
-		// Remove one or more patterns from the end, optionally followed by a dot and wrapped in parentheses; multiple patterns are only removed if:
-		//  - the first patterns ends with a dot or a closing parenthesis, optionally followed by a comma, or,
-		//  - the first pattern is followed by an ampersand or a hyphen.
-		return new RegExp(`\\s+(?:\\(?(?:${patterns.join('|')})\\.?\\)?(?:(?<=[.)]),?\\s*|\\s*[&-]\\s*|$))+$`, 'i');
-	});
-}
-
-function buildPrefixPatterns (names: string[], abbrs: string[]) {
-	return buildPatterns(names, abbrs, (patterns: string[]) => {
-		return new RegExp(`^(?:${patterns.join('|')})[.,]?\\s+`, 'i');
-	});
-}
-
-function buildPatterns (names: string[], abbrs: string[], builder: (patterns: string[]) => RegExp) {
-	const preparedNames: string[] = [];
-	const preparedAbbrs: string[] = [];
-
-	for (const name of names) {
-		const namePattern = _.escapeRegExp(name)
-			// Add a dot to the end of each word.
-			.replace(/([^.])(?:\s+|$)/g, '$1\\.')
-			// Remove whitespace after dots.
-			.replace(/\.\s+/g, '.')
-			// Allow whitespace after dots, make all existing dots replaceable by whitespace.
-			.replace(/\\\.(?=.)/g, '[. ] *')
-			// Remove the final dot.
-			.replace(/\\\.$/, '');
-
-		preparedNames.push(namePattern);
-	}
-
-	for (const abbr of abbrs) {
-		const abbrPattern = _.escapeRegExp(abbr)
-			// Add a dot to the end of each word.
-			.replace(/([^.])(?:\s+|$)/g, '$1\\.')
-			// Remove whitespace after dots.
-			.replace(/\.\s+/g, '.')
-			// Allow whitespace after dots, make all existing dots optional.
-			.replace(/\\\.(?=.)/g, '\\.? *')
-			// Allow optional dots or whitespace between any two word characters
-			.replace(/(?<=\w)(?=\w)/g, '[. ]*')
-			// Remove the final dot.
-			.replace(/\\\.$/, '');
-
-		preparedAbbrs.push(abbrPattern);
-	}
-
-	return {
-		namesPattern: builder(preparedNames),
-		abbrsPattern: builder(preparedAbbrs),
-	};
-}
 
 async function collectLegalForms (legalFormsData: CsvLegalFormRow[], minSynthesizedAbbreviationLength: number = 2) {
 	const legalFormsName = new Set<string>();
@@ -273,13 +188,7 @@ async function collectLegalForms (legalFormsData: CsvLegalFormRow[], minSynthesi
 		abbr.forEach(a => legalFormsAbbr.add(ascii(a).toLowerCase().replace(/[., ()]/g, '')));
 	});
 
-	// Convert to array and sort by length (longest first) to avoid partial matches
-	return {
-		names: Array.from(legalFormsName)
-			.sort((a, b) => b.length - a.length),
-		abbrs: Array.from(legalFormsAbbr)
-			.sort((a, b) => b.length - a.length),
-	};
+	return { names: legalFormsName, abbrs: legalFormsAbbr };
 }
 
 const readLegalFormsFile = () => new Promise<{ allForms: CsvLegalFormRow[]; prefixForms: CsvLegalFormRow[] }>((resolve, reject) => {