--- a/test/tests/unit/geoip/legal-name-normalization.test.ts
+++ b/test/tests/unit/geoip/legal-name-normalization.test.ts
@@ -28,6 +28,35 @@ describe('legal-name-normalization', () => {
 
 		// Already normalized or no legal suffix
 		{ original: 'AkileCloud Network', expected: 'AkileCloud Network' },
+
+		{ original: 'Hangzhou Alibaba Advertising Co.,Ltd.', expected: 'Hangzhou Alibaba Advertising' },
+		{ original: 'Shanghai Mobile Communications Co.,Ltd.', expected: 'Shanghai Mobile Communications' },
+		{ original: 'Alibaba (US) Technology Co., Ltd.', expected: 'Alibaba (US) Technology' },
+		{ original: 'Siamdata Communication Co.,Ltd.', expected: 'Siamdata Communication' },
+		{ original: 'China Mobile Communications Group Co., Ltd.', expected: 'China Mobile Communications Group' },
+		{ original: 'Rackzar  (Pty) Ltd', expected: 'Rackzar' },
+		{ original: 'Henan Mobile Communications Co. Ltd', expected: 'Henan Mobile Communications' },
+		{ original: 'Web Squad Connect (Pty) Ltd', expected: 'Web Squad Connect' },
+		{ original: 'Web Dadeh Paydar Co (Ltd)', expected: 'Web Dadeh Paydar Co' },
+		{ original: 'Jinx Co. Limited', expected: 'Jinx' },
+
+		{ original: 'DA International Group Ltd.', expected: 'DA International Group' },
+		{ original: 'OKB PROGRESS LLC', expected: 'OKB PROGRESS' },
+		{ original: 'Limited Company Information and Consulting Agency', expected: 'Limited Company Information and Consulting Agency' },
+		{ original: 'IP Vendetta Inc.', expected: 'IP Vendetta' },
+		{ original: 'IP ServerOne Solutions Sdn Bhd', expected: 'IP ServerOne Solutions' },
+		{ original: 'PT. Elektrindo Data Nusantara', expected: 'Elektrindo Data Nusantara' },
+
+		{ original: 'Microchip s.c. W. Wrodarczyk, A. Kossowski', expected: 'Microchip s.c. W. Wrodarczyk, A. Kossowski' },
+		{ original: 'Tencent Building, Kejizhongyi Avenue', expected: 'Tencent Building, Kejizhongyi Avenue' },
+		{ original: 'Hollander & Jacobsen Ug (Haftungsbeschraenkt)', expected: 'Hollander & Jacobsen' },
+		{ original: 'eServer s.r.o.', expected: 'eServer' },
+		{ original: 'synlinq.de', expected: 'synlinq.de' },
+		{ original: 'ServerPoint.com', expected: 'ServerPoint.com' },
+		{ original: 'S.C. INFOTECH-GRUP S.R.L.', expected: 'S.C. INFOTECH-GRUP' },
+		{ original: 'S.SETEVAYA SVYAZ, OOO', expected: 'S.SETEVAYA SVYAZ' },
+		{ original: 'TELECOMUNICACOES ALARCAO E FERNANDES LTDA - ME', expected: 'TELECOMUNICACOES ALARCAO E FERNANDES' },
+		{ original: 'M & R NETWORK LTDA - ME', expected: 'M & R NETWORK' },
 	];
 
 	for (const { original, expected } of cases) {