- result is exactly the same as with regexes for all current probes
- speed with node v22.16.0:
```
tsx normalize-from-code.js probes.json
normalizeLegalName xTom GmbH: 0.123ms
normalizeLegalName STYLISH BY A&L SRL: 0.059ms
normalizeLegalName HostPapa: 0.007ms
normalizeLegalName HostPapa: 0.003ms
normalizeLegalName Odido Netherlands B.V.: 0.01ms
normalizeLegalName Odido Netherlands B.V.: 0.008ms
normalizeLegalName Eons Data Communications Limited: 0.068ms
normalizeLegalName AkileCloud Network: 0.011ms
normalizeLegalName GLOBAL INTERNET SOLUTIONS LLC: 0.009ms
normalizeLegalName Oracle Corporation: 0.004ms
normalizeLegalName Terabix: 0.006ms
normalizeLegalName Hetzner Online GmbH: 0.004ms
...
```